requirements.txt
Flask==2.3.2
qrcode[pil]==7.4.2
Pillow==9.5.0